01_Chatbot/requirements.txt
langchain
langchain-openai  # Or langchain if you are using an older version
streamlit
python-dotenv